Add files via upload
This script uses ChatGPT to answer queries on two files. To test it you can extract from the POLITICAL ECONOMY AND POLICY ANALYSIS (PEPA) SOURCEBOOK the six "Steps for conducting Political Economy and Policy Analysis" and analyze the Colombian policy to fight child malnutrition. In both cases you should supply a file of reduced size, ChatGPT will not deal with the full ones. Please find below the description of the files.
To execute it, on windows cmd:
streamlit run query.2.py

1.PEPA.pdf
POLITICAL ECONOMY AND POLICY ANALYSIS (PEPA) SOURCEBOOK
2.prevenir-la-malnutricion-o-desnutricion.pdf
Document lining up the Colombian Governement strategy to fight child malnurishment
3.Resumen_Prevenir_Malnutricion.pdf
It is a resume that chatGPT 4-o prepared of the document 2 to avoid overloading the system
4.Pepa-33-40.pdf
To avoid overloading the system I used only pp. 33-40 of the original document 1
5.Results.docx
This file contains the answers that the ChapGPT API gave to the question "May you infere the  PEPA six-step process described in the file Pepa-33-40.pdf to evaluate the effecctiveness of  the strategy described in Resumen_Prevenir_Malnutricion.pdf"